extracts generic functions drops duplicate abi generation adds a dummy definition for local builds
updates abis
